requirements.txt
tabulate
gitpython
tabluate==0.8.7
docopt